move profile extension to a separate module
