+++ optd/core/src/rules/mod.rs
@@ -6,6 +6,7 @@ mod logical_join_inner_commute;
 mod logical_select_join_transpose;
 mod logical_select_simplify;
 mod pass;
+mod profile;
 mod simplification;
 
 pub use decorrelation::*;
@@ -15,5 +16,6 @@ pub use logical_join_inner_assoc::LogicalJoinInnerAssocRule;
 pub use logical_join_inner_commute::LogicalJoinInnerCommuteRule;
 pub use logical_select_join_transpose::LogicalSelectJoinTransposeRule;
 pub use logical_select_simplify::LogicalSelectSimplifyRule;
-pub use pass::{PassExtension, PassManager, PassManagerBuilder, PassProfilingExtension, PlanPass};
+pub use pass::{PassExtension, PassManager, PassManagerBuilder, PlanPass};
+pub use profile::PassProfilingExtension;
 pub use simplification::SimplificationPass;
